restructure project wave 6

+++ backend/ZiziBot.Common/Dtos/TelegramSinkConfigDto.cs
@@ -4,5 +4,5 @@ public class TelegramSinkConfigDto
 {
     public string? BotToken { get; init; }
     public long? ChatId { get; init; }
-    public long? ThreadId { get; set; }
+    public int? ThreadId { get; set; }
 }
